requirements.txt
requests
mysql-connector-python
python-dotenv
Flask
Werkzeug
openai